+++ compiler/src/iree/compiler/Dialect/LinalgExt/IR/LinalgExtOps.cpp
@@ -43,8 +43,6 @@
 #include <cstdint>
 #include <optional>
 
-#define DEBUG_TYPE "linalg-ext-ops"
-
 namespace mlir::iree_compiler::IREE::LinalgExt {
 
 //===----------------------------------------------------------------------===//
@@ -1808,329 +1806,6 @@ CustomOp::reifyResultShapes(OpBuilder &builder,
         tensor::getMixedSizes(builder, getLoc(), init);
     reifiedReturnShapes.emplace_back(std::move(sizes));
   }
-  return success();
-}
-
-//===---------------------------------------------------------------------===//
-// Implementation of `TilingInterface`
-//===---------------------------------------------------------------------===//
-
-/// These methods copied/modified from `TilingInterface` implementation of
-/// `getIterationDomain` of `LinalgOp`s.
-
-static SmallVector<OpFoldResult>
-createFlatListOfOperandDims(OpBuilder &builder, Location loc,
-                            CustomOp customOp) {
-  SmallVector<OpFoldResult> result;
-  for (Value operand : customOp->getOperands()) {
-    for (auto dim : llvm::seq<unsigned>(customOp.getRank(operand))) {
-      result.push_back(linalg::createFoldedDimOp(builder, loc, operand, dim));
-    }
-  }
-  return result;
-}
-
-/// Return dim expresssions that can be used as replacements in map that
-/// contains `numSymbols` symbols. The new dim expressions have positions
-/// `numDims, numDims + 1, numDims + 2, ...., numDims + numSymbols - 1`.
-static SmallVector<AffineExpr> getDimExprsForSymbols(MLIRContext *context,
-                                                     unsigned numDims,
-                                                     unsigned numSymbols) {
-  return llvm::map_to_vector(
-      llvm::seq<unsigned>(0, numSymbols), [&](unsigned symbolNumber) {
-        return getAffineDimExpr(symbolNumber + numDims, context);
-      });
-}
-
-/// Convert all symbols in the map to dim expressions, such that the new dim
-/// expressions have positions `numDims, numDims + 1, numDims + 2, ...., numDims
-/// + numSymbols - 1`.
-static AffineMap
-convertDimsToSymbols(AffineMap map, unsigned numDims, unsigned numSymbols,
-                     SmallVector<AffineExpr> &symbolReplacements) {
-  return map.replaceDimsAndSymbols(/*dimReplacements=*/ArrayRef<AffineExpr>{},
-                                   symbolReplacements, numDims + numSymbols, 0);
-}
-static SmallVector<AffineMap>
-convertDimsToSymbols(ArrayRef<AffineMap> maps, unsigned numDims,
-                     unsigned numSymbols,
-                     SmallVector<AffineExpr> &symbolReplacements) {
-  return llvm::map_to_vector(maps, [&](AffineMap map) {
-    return convertDimsToSymbols(map, numDims, numSymbols, symbolReplacements);
-  });
-}
-static SmallVector<AffineMap> convertDimsToSymbols(MLIRContext *context,
-                                                   ArrayRef<AffineMap> maps,
-                                                   unsigned numDims,
-                                                   unsigned numSymbols) {
-  auto symbolReplacements = getDimExprsForSymbols(context, numDims, numSymbols);
-  return convertDimsToSymbols(maps, numDims, numSymbols, symbolReplacements);
-}
-
-SmallVector<Range> CustomOp::getIterationDomainForDimensions(
-    OpBuilder &builder, ArrayRef<unsigned> dims, ArrayRef<unsigned> symbols) {
-  CustomOp customOp = *this;
-  SmallVector<AffineMap> maps = customOp.getIndexingMapsArray();
-  if (maps.empty()) {
-    return SmallVector<Range>{};
-  }
-
-  Location loc = getLoc();
-
-  // 1. Create a flat list of all the operand shapes (similar to Linalg)
-  SmallVector<OpFoldResult> allShapesSizes =
-      createFlatListOfOperandDims(builder, loc, customOp);
-
-  // 2a. Next we need to get a map from shapes to loop. Since `custom_op`
-  //     has indexing maps that have symbols, to make this work correctly
-  //     compute new maps that replaces the symbols with "new" dims.
-  unsigned numDims = getNumLoops();
-  unsigned numSymbols = getNumNonLoopDimensions();
-  MLIRContext *context = getContext();
-  SmallVector<AffineMap> modifiedMaps =
-      convertDimsToSymbols(context, maps, numDims, numSymbols);
-
-  // 2b. Concat the affine maps.
-  AffineMap concatMap = inversePermutation(concatAffineMaps(modifiedMaps));
-  SmallVector<Range> ranges;
-
-  OpFoldResult zero = builder.getIndexAttr(0), one = builder.getIndexAttr(1);
-  auto getRange = [&](AffineExpr expr) {
-    OpFoldResult ofr = affine::makeComposedFoldedAffineApply(builder, loc, expr,
-                                                             allShapesSizes);
-    return Range{zero, ofr, one};
-  };
-  ranges = llvm::map_to_vector(
-      dims, [&](unsigned dim) { return getRange(concatMap.getResult(dim)); });
-  ranges.append(llvm::map_to_vector(symbols, [&](unsigned symbol) {
-    return getRange(concatMap.getResult(symbol + numSymbols));
-  }));
-  return ranges;
-}
-
-SmallVector<Range> CustomOp::getIterationDomain(OpBuilder &builder) {
-  auto dims = llvm::to_vector(llvm::seq<unsigned>(0, getNumLoops()));
-  return getIterationDomainForDimensions(builder, dims,
-                                         /*symbols=*/ArrayRef<unsigned>{});
-}
-
-/// During tiling, the tiling implementation generates offsets and sizes
-/// to use for the dimensions of the operation that correspond to loops.
-/// The dimensions of the operation corresponding to symbols isnt tiled.
-/// This method extends the offsets and sizes for a tile with the
-/// offsets (which are zero) and sizes (which are same as the untiled sizes) for
-/// the dimension represented by symbols. The
-static std::pair<SmallVector<OpFoldResult>, SmallVector<OpFoldResult>>
-appendOffsetAndSizeForSymbolDimensions(OpBuilder &builder, CustomOp customOp,
-                                       ArrayRef<OpFoldResult> offsets,
-                                       ArrayRef<OpFoldResult> sizes) {
-  unsigned numSymbols = customOp.getNumNonLoopDimensions();
-  if (numSymbols == 0) {
-    return {llvm::to_vector(offsets), llvm::to_vector(sizes)};
-  }
-
-  auto appendedOffsets = llvm::to_vector(offsets);
-  appendedOffsets.resize(appendedOffsets.size() + numSymbols,
-                         builder.getIndexAttr(0));
-
-  auto appendedSizes = llvm::to_vector(sizes);
-  auto symbols = llvm::to_vector(llvm::seq<unsigned>(0, numSymbols));
-  auto symbolRanges = customOp.getIterationDomainForDimensions(
-      builder, /*dims=*/ArrayRef<unsigned>{}, symbols);
-  auto symbolRangesUb =
-      llvm::map_range(symbolRanges, [](Range r) { return r.size; });
-  appendedSizes.append(symbolRangesUb.begin(), symbolRangesUb.end());
-  return {appendedOffsets, appendedSizes};
-}
-
-/// This method is adapted from the method `linalg::computeAllSliceParameters`,
-/// adapted to work with map that have symbols, and empty maps.
-SmallVector<std::optional<linalg::SliceParameters>>
-computeCustomOpAllSliceParameters(OpBuilder &builder, Location loc,
-                                  CustomOp customOp, ValueRange valuesToTile,
-                                  ArrayRef<OpFoldResult> ivs,
-                                  SmallVector<OpFoldResult> tileSizes) {
-  assert(ivs.size() == static_cast<size_t>(llvm::count_if(
-                           llvm::make_range(tileSizes.begin(), tileSizes.end()),
-                           [](OpFoldResult v) { return !isZeroIndex(v); })) &&
-         "expected as many ivs as non-zero sizes");
-  unsigned numDims = customOp.getNumLoops();
-  unsigned numSymbols = customOp.getNumNonLoopDimensions();
-
-  // Construct (potentially temporary) mins and maxes on which to apply maps
-  // that define tile subshapes.
-  SmallVector<OpFoldResult> lbs =
-      linalg::computeTileOffsets(builder, loc, ivs, tileSizes);
-  SmallVector<OpFoldResult> sizeBounds = llvm::to_vector(tileSizes);
-
-  std::tie(lbs, sizeBounds) = appendOffsetAndSizeForSymbolDimensions(
-      builder, customOp, lbs, sizeBounds);
-
-  tileSizes.resize(tileSizes.size() + numSymbols, builder.getIndexAttr(0));
-  SmallVector<OpFoldResult> subShapeSizes =
-      linalg::computeTileSizes(builder, loc, tileSizes, sizeBounds);
-
-  SmallVector<AffineExpr> symbolReplacements;
-  if (numSymbols != 0) {
-    symbolReplacements =
-        getDimExprsForSymbols(builder.getContext(), numDims, numSymbols);
-  }
-
-  assert(static_cast<int64_t>(valuesToTile.size()) <=
-             customOp->getNumOperands() &&
-         "more value to tile than operands.");
-  SmallVector<std::optional<linalg::SliceParameters>> allSliceParams;
-  allSliceParams.reserve(valuesToTile.size());
-  for (auto [opOperand, val] :
-       llvm::zip(customOp->getOpOperands(), valuesToTile)) {
-    Value shapedOp = val;
-    LLVM_DEBUG(llvm::dbgs() << "makeTiledShapes: for operand " << shapedOp);
-    AffineMap map = customOp.getMatchingIndexingMap(&opOperand);
-    if (numSymbols != 0) {
-      map = convertDimsToSymbols(map, numDims, numSymbols, symbolReplacements);
-    }
-
-    // If the map is empty, we dont tile this operand.
-    Type operandType = opOperand.get().getType();
-    if (map.isEmpty() || !isa<ShapedType>(operandType)) {
-      allSliceParams.push_back(std::nullopt);
-      LLVM_DEBUG(llvm::dbgs()
-                 << ": not tiled: use shape: " << operandType << "\n");
-      continue;
-    }
-
-    allSliceParams.push_back(linalg::computeSliceParameters(
-        builder, loc, shapedOp, tileSizes, map, lbs, sizeBounds, subShapeSizes,
-        /*omitPartialTileCheck=*/true));
-  }
-
-  return allSliceParams;
-}
-
-/// This method is same as `materializeTiledShape` method defined in
-/// `mlir/Dialect/Linalg/Utils/Utils.[h|cpp]`.
-static Operation *
-materializeTiledShape(OpBuilder &builder, Location loc, Value valueToTile,
-                      const linalg::SliceParameters &sliceParams) {
-  auto shapedType = dyn_cast<ShapedType>(valueToTile.getType());
-  auto *sliceOp = TypeSwitch<ShapedType, Operation *>(shapedType)
-                      .Case([&](RankedTensorType) {
-                        return builder.create<tensor::ExtractSliceOp>(
-                            loc, valueToTile, sliceParams.offsets,
-                            sliceParams.sizes, sliceParams.strides);
-                      })
-                      .Default([](ShapedType) -> Operation * {
-                        llvm_unreachable("Unexpected shaped type");
-                      });
-  return sliceOp;
-}
-
-/// This method is adapted from the method `linalg::makeTiledShapes`,
-/// adapted to work with map that have symbols, and empty maps.
-SmallVector<Value> makeCustomOpTiledShapes(OpBuilder &builder, Location loc,
-                                           CustomOp customOp,
-                                           ValueRange valuesToTile,
-                                           ArrayRef<OpFoldResult> ivs,
-                                           ArrayRef<OpFoldResult> tileSizes) {
-  SmallVector<std::optional<linalg::SliceParameters>> allSliceParameter =
-      computeCustomOpAllSliceParameters(builder, loc, customOp, valuesToTile,
-                                        ivs, llvm::to_vector(tileSizes));
-  SmallVector<Value> tiledShapes;
-  for (auto item : llvm::zip(valuesToTile, allSliceParameter)) {
-    Value valueToTile = std::get<0>(item);
-    std::optional<linalg::SliceParameters> sliceParams = std::get<1>(item);
-    tiledShapes.push_back(
-        sliceParams.has_value()
-            ? materializeTiledShape(builder, loc, valueToTile, *sliceParams)
-                  ->getResult(0)
-            : valueToTile);
-  }
-  return tiledShapes;
-}
-
-static void offsetCustomOpIndices(OpBuilder &b, CustomOp customOp,
-                                  ArrayRef<OpFoldResult> offsets) {
-  IRRewriter rewriter(b);
-  for (auto indexOp : customOp.getBody()->getOps<IREE::LinalgExt::IndexOp>()) {
-    if (indexOp.getDim() >= offsets.size() || !offsets[indexOp.getDim()])
-      continue;
-    OpBuilder::InsertionGuard guard(b);
-    rewriter.setInsertionPointAfter(indexOp);
-    AffineExpr index, offset;
-    bindDims(b.getContext(), index, offset);
-    OpFoldResult applied = affine::makeComposedFoldedAffineApply(
-        rewriter, indexOp.getLoc(), index + offset,
-        {getAsOpFoldResult(indexOp.getResult()), offsets[indexOp.getDim()]});
-    Value materialized =
-        getValueOrCreateConstantIndexOp(b, indexOp.getLoc(), applied);
-    rewriter.replaceUsesWithIf(indexOp, materialized, [&](OpOperand &use) {
-      return use.getOwner() != materialized.getDefiningOp();
-    });
-  }
-}
-
-FailureOr<TilingResult>
-CustomOp::getTiledImplementation(OpBuilder &builder,
-                                 ArrayRef<OpFoldResult> offsets,
-                                 ArrayRef<OpFoldResult> sizes) {
-  CustomOp customOp = *this;
-  // Leave the `sizeBounds` value empty. That is only needed when the `sizes`
-  // specified could lead to out of bounds accesses.
-  Location loc = getLoc();
-
-  SmallVector<Value> valuesToTile = getOperation()->getOperands();
-  SmallVector<Value> tiledOperands = makeCustomOpTiledShapes(
-      builder, loc, *this, valuesToTile, offsets, sizes);
-  SmallVector<Operation *> generatedSlices = llvm::map_to_vector(
-      llvm::make_filter_range(
-          tiledOperands,
-          [](Value v) -> bool {
-            return isa_and_nonnull<tensor::ExtractSliceOp, memref::SubViewOp>(
-                v.getDefiningOp());
-          }),
-      [](Value v) -> Operation * { return v.getDefiningOp(); });
-
-  SmallVector<Type> resultTensorTypes =
-      llvm::map_to_vector(getDpsInitsMutable(), [&](OpOperand &opOperand) {
-        return tiledOperands[opOperand.getOperandNumber()].getType();
-      });
-
-  Operation *tiledOp =
-      mlir::clone(builder, customOp, resultTensorTypes, tiledOperands);
-  offsetCustomOpIndices(builder, cast<CustomOp>(tiledOp), offsets);
-
-  return TilingResult{
-      {tiledOp}, SmallVector<Value>(tiledOp->getResults()), generatedSlices};
-}
-
-/// Methods copied/modified from `TilingInterface` implementation of
-/// `getTiledImplementation` of `LinalgOp`s.
-
-LogicalResult CustomOp::getResultTilePosition(
-    OpBuilder &builder, unsigned resultNumber, ArrayRef<OpFoldResult> offsets,
-    ArrayRef<OpFoldResult> sizes, SmallVector<OpFoldResult> &resultOffsets,
-    SmallVector<OpFoldResult> &resultSizes) {
-  CustomOp customOp = *this;
-  Location loc = getLoc();
-
-  AffineExpr d0;
-  bindDims(builder.getContext(), d0);
-  SmallVector<OpFoldResult> appendedOffsets, appendedSizes;
-  std::tie(appendedOffsets, appendedSizes) =
-      appendOffsetAndSizeForSymbolDimensions(builder, customOp, offsets, sizes);
-
-  SmallVector<OpFoldResult> subShapeSizes =
-      llvm::to_vector(llvm::map_range(appendedSizes, [&](OpFoldResult ofr) {
-        return affine::makeComposedFoldedAffineApply(builder, loc, d0 - 1, ofr);
-      }));
-
-  OpOperand *outOperand = customOp.getDpsInitOperand(resultNumber);
-  linalg::SliceParameters sliceParams = linalg::computeSliceParameters(
-      builder, loc, outOperand->get(), sizes,
-      customOp.getMatchingIndexingMap(outOperand), appendedOffsets,
-      /*ubs*/ {}, subShapeSizes, /*omitPartialTileCheck=*/true);
-  resultOffsets = sliceParams.offsets;
-  resultSizes = sliceParams.sizes;
   return success();
 }
